requirements.txt
fastapi~=0.85.0
uvicorn
jinja2
numpy
scikit-learn
python-dotenv
sqlalchemy
pg8000
cloud-sql-python-connector
alembic
argon2-cffi
python-multipart
starlette
itsdangerous